Consolidando as chamadas de usuário no serviço dele
